feat: 장바구니 삭제 시 confirm

+++ src/components/cart/CartItem/CartItem.tsx
@@ -24,7 +24,9 @@ function CartItem({ cart }) {
   };
 
   const onClickDeleteButton = async () => {
-    dispatch(fetchDeleteCartAsync(id) as any);
+    if (confirm('정말로 삭제하시겠습니까?')) {
+      dispatch(fetchDeleteCartAsync(id) as any);
+    }
   };
 
   const onClickIncreaseButton = () => {
